[REFACTOR] UserId 관련 클래스 user 패키지로 이동
