+++ src/main/kotlin/io/gijung/aichatbot/auth/application/port/in/SignUpUseCase.kt
@@ -1,6 +1,6 @@
 package io.gijung.aichatbot.auth.application.port.`in`
 
-import io.gijung.aichatbot.auth.domain.model.vo.UserId
+import io.gijung.aichatbot.user.domain.modle.vo.UserId
 
 fun interface SignUpUseCase {
     fun signUp(command: SignUpCommand): UserId
