+++ src/main/kotlin/io/gijung/aichatbot/auth/infrastructure/id/UuidGenerator.kt
@@ -1,7 +1,7 @@
 package io.gijung.aichatbot.auth.infrastructure.id
 
 import io.gijung.aichatbot.auth.application.port.out.UserIdGenerator
-import io.gijung.aichatbot.auth.domain.model.vo.UserId
+import io.gijung.aichatbot.user.domain.modle.vo.UserId
 import org.springframework.stereotype.Component
 import java.util.*
